requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
pyyaml

sec-edgar-downloader
beautifulsoup4

chromadb
sentence-transformers
unstructured 
pypdf 

streamlit